Update patch set 6
Patch Set 6:

copying Sandrine's comment 
"
Line 223:
I would prefer you move these source files into each image's Makefile and leave the list of files in this Makefile as is. This would be more consistent with what we've done so far. It helps keeping the top Makefile as clean as possible, as well as reducing the dependency of test images on each other, if we ever need to add some extra PAuth file for some specific test image some day.

So for example, for TFTF I suggest we add these 2 PAuth source files in the FRAMEWORK_SOURCES variable inside the TFTF makefile and then line 228 does not need any changes."

Patch-set: 6