doc: Copy readme content to docs/index.rst
The docs/index.rst file is the new landing page / entry point
for reading the docs so the top-level content needs to be copied
there.

The existing content in readme.rst cannot be modified to point to
the new page until the hosted copy of the docs has gone live (in
order to have a working URL).

Change-Id: Ifa9e65d463c2887bbd53f7224aa803796aeec68f
Signed-off-by: Paul Beesley <paul.beesley@arm.com>