plat: nvidia: introduce platform port for Tegra194
This patch contains the basic platform support for Tegra194 platforms
to initialize the tftf framework and execute tests on the CPUs.

The tests require support for a timer, non-volatile memory, UART and
GIC. This port uses Tegra_TMR0 as the timer, TI UART 16550 as the
console and DRAM as NVM. The GIC driver provided by ARM is used as is.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I2da2121cb05445f47b7d3083cdd6ac6d52586797