doc: Migrate content from User Guide and update it
The User Guide document holds many different bits of
information that are not always related to each other. These
are moved out into files where information is grouped tightly
together.

This also updates build instructions and environment variable
instructions to link to the relevent parts of the TF-A user guide
and use more recent versions of compiler, respetively.

Change-Id: Iace851f93c7dea8f154ae52a2665dd599f2e5d72
Signed-off-by: Paul Beesley <paul.beesley@arm.com>
Signed-off-by: Jimmy Brisson <jimmy.brisson@arm.com>