doc: Update link formats
Use docutils/Sphinx-type linking with the :ref: keyword where
possible, instead of direct links to files.

As part of this patch some external links have been updated too,
especially those that point to the TF-A docs. Some link targets
are no longer valid but can be restored after the TF-A User Guide
is split up.

Change-Id: I328d5141cf9b32471bf669bfc0191e8f6f883a48
Signed-off-by: Paul Beesley <paul.beesley@arm.com>
Signed-off-by: Jimmy Brisson <jimmy.brisson@arm.com>