timers: remove dependency on SYS_CNT_BASE1
The ARM ARM Section D11.1.1 titled "The full set of Generic Timer components"
says that Memory-mapped timers are optional. The timer framework and tests use
the SYS_CNT_BASE1 macro to read the memory-mapped timers, but they can also
ready the system registers to get the same values, to accomodate all ARM
platforms.

This patch removes the usage of SYS_CNT_BASE1 from the timer code and replaces
it with calls to read_cntpct_el0().

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I4d9ee5587351ee737800539f4d48606e7de80538