Update patch set 1
Patch Set 1: Code-Review+1

As far as 16550 driver code is concerned it looks good to me.
But an open question is, till now in tf-a-tests only arm/pl011 driver was used so there was no real usecase for having console framework similar to that of tf-a where any uart driver can be registered. Do you think is it worth replicating console framework here also? or is it an overkill ??

Patch-set: 1
Reviewer: Gerrit User 1000102 <1000102@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>
Label: Code-Review=+1