Query for invalid MPIDS when populating the power domain tree.
Currently, TFTF marks all the possible CPU power domains as
available regardless of whether they actually exist or not.
This can lead to problems if the platform has holes on the
topology description.

To address that, this patch queries the platform about any
possible MPIDR, marking its CPU power domain as present only
if its MPIDR is valid.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I1570088099cd5ff1ff2f33add1f76e5ad8d200b5