spm: introduce helper function to detect SEL2 presence
This patch introduces a helper function to detect if we are running
on a CPU that supports SEL2 exception mode.

For pre-8.4 CPUs, this check helps the sp_helpers return '0' for number
of VCPUs and VMs on a physical CPU. The HVC calls are skipped for pre-8.4
CPUs too.

Verified by booting Cactus on a Tegra194 (pre-8.4) platform.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I43a6d4151e47aaab556a75e2b3cdade9794e0480