Tegra194: enable "spe-uart" console
This patch enables the UART functionality provided by the
SPE core. The SPE core is the arbiter and all clients must
communicate with SPE instead of writing to the UART directly.
This ensures that the prints on the UART console do not
appear garbled due to multiple client writes at the same time.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: If40b13bb9bfb5a6d505cacec211b1aa931adc268