Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 3: Code-Review+1
> 
> Looks good to me. So Varun's suggested fix worked? what was the real issue?

As Varun and other folks suggested, the large delay was not needed. I talked to Arm support engineer and found that actually no special delay is required since the write is a store operation to a memory region mapped as Device Memory. 
Based on Varun's suggestion, I added a while loop as extra precaution. But it helped to catch a small issue which it introduced itself. The mask 0xFF is not appropriate because JUNO only has 6 cores. So the desired mask to request the GIC to forward SPI to all CPUs would be 0x3F(binary one-hot encoded: 0011_1111) though writing 0xFF doesn't hurt. When read, it would return the value 0x3F.

With these changes, the test still doesn't pass but fails cleanly. The bug is actually in TF-A which I will be working to fix shortly.

Patch-set: 4