spm: introduce support for Cactus as SPMC
This patch introduces a new build flag, CACTUS_SEL1_SPMC, to allow
Cactus to run as SPMC and enable the following

* assume number of VMs as '1'
* assume number of VCPUs as '1'
* assume VM ID as '1'
* skips partition_info_get boot tests

The spm helper functions for Cactus have been moved to Cactus to
allow this change to compile for 'cactus' and 'cactus_mm' targets.

The platforms using Cactus in this setting, need to enable the flag
from their makefiles.

Verified by booting Cactus as SPMC on a Tegra194 platform.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: Id453afdd3959899635fdb2eebdef70bbe186e3e2