cactus: IRQ handling with managed exit mechanism
While executing in secure world, NS interrupts are trapped in SPM(S-EL2)
as FIQ, which then injects a managed exit virtual IRQ to running Secure
Partition, so that it can save its context and return to normal world
for actual handling of interrupt.

This patch implements Cactus IRQ handler, which sends INTERRUPT_GET
hypervisor call to get the interrupt number and in case of managed exit
it saves its context (empty as of now) and sends MSG_SEND_DIRECT_RESP
ABI call to SPMC for interrupt to be handled by normal world. For any
other interrupt it panics.

Change-Id: I65e42645330be0787449be850579e1a9fa35127b
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>