Cactus: Commands to test bind and unbind notifications
Added test command to test FFA_NOTIFICATION_BIND and
FFA_NOTIFICATION_UNBIND interfaces.
Both commands allow for the arguments to the call to be propagated on
the request payload. If the call to the interfaces returns with an error
Cactus places on the response CACTUS_ERROR and adds the respective
FF-A error code.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I3fc3a5e559d0c6d6742f5d86973f6e67891ad790