feat(cactus): commands for notifications bindings
Added test command to test FFA_NOTIFICATION_BIND and
FFA_NOTIFICATION_UNBIND interfaces.
Both commands allow for the arguments to the call to be propagated in
the request payload. If the call to the interfaces returns with an error
Cactus places in the response CACTUS_ERROR and adds the respective
FF-A error code.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I3fc3a5e559d0c6d6742f5d86973f6e67891ad790