Cactus: Commands to test set and get notifications
Added test command to test FFA_NOTIFICATION_SET and
FFA_NOTIFICATION_GET interfaces.
Both commands allow for the arguments to the call to be propagated on
the request payload. If the call to the interfaces returns with an error
Cactus responds with CACTUS_ERROR and adds the respective FF-A error
code.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I7c3425bf240cafd89613e4c5cc2ece6d58625ac2