test(cactus): prevent realm region access from swd
This change adds TFTF and cactus test to permit checking a realm region
cannot be accessed from secure world.
A non-secure buffer is delegated to realm and shared to a secure
partition. The SP retrieves the region from the SPM, maps it and
attempts a read access to the region. It is expected to trigger a GPF
data abort on the PE caught by a custom exception handler.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I8f855f394d0490b3584e60ceba4f3d2a20197495