test(ff-a): FFA_NOTIFICATION_INFO_GET interface
Simple test to FFA_NOTIFICATION_INFO_GET. Added helper functions
to call the interface, and validate its return. Changed slightly the
tests to signal a notification from NWd to an SP, and from SP to a VM,
using the referred test functions. This validates the simplest return
scenario for the FFA_NOTIFICATION_INFO_GET interface.

Change-Id: I4fdc9c00b30244e6a0caad668a006c6a825c8026
Signed-off-by: J-Alves <joao.alves@arm.com>