smc(fuzzer): random seed generation and build parameters
This patch adds build parameters to TFTF to allow more control of the
SMC fuzzing test suite as well as randomly generating seeds for the
pseudo-random number generator used for SMC fuzzing tests.

Build Options

SMC_FUZZ_CALLS_PER_INSTANCE=n: Sets the number of SMC calls per test
instance, where n is the number of calls. If not provided, this value
defaults to 100.

SMC_FUZZ_INSTANCE_COUNT=n: Sets the number of instances to run, where n
is the number of instances. If not provided, this value defaults to 1.

SMC_FUZZ_SEEDS=i,j,k,...: If randomly generated seeds are not desired,
such as a case where a certain seed results in a fail, seeds can be
provided in the build command.  The seeds can either be integers or hex
values, must be formatted as shown with a comma between each seed and
no spaces, and the number of seeds provided must match
SMC_FUZZ_INSTANCE_COUNT. If not provided, seeds are randomly generated
by the build system.

Signed-off-by: John Powell <john.powell@arm.com>
Change-Id: I2cc6f4d7a02ec4b590db335ac098c9d41df42f0c