test(ff-a): signaling per-vCPU notifications
Testing per vCPU notifications, which are notifications whose vCPU is
specified by the sender. Each vCPU is targeted with a notification, as
such all CPUs are brought up, in each one of them TFTF sends a request
to the receiver Cactus SP to get its pending notifications.
Small changes to helper functions to include the flags argument to be
passed to notifications FF-A calls. Tests encompass SP to SP and NWd to
SP.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I86ba840b4c1d35cce95e2bc197293511db6783a1