refactor(cactus): notification set command for SRI test
Refactored slightly the notification set command in order to validate
behavior from SWd for different configurations of the SRI flag.
The test command now accepts another argument, which specifies the ID
of an SP to send a request after setting the notification. This will be
useful to test that sending SRI was delayed, and after the notification
set call the execution continued in the SWd. TFTF will be able to check
if the specified SP processed a command recurring to the test command
CACTUS_GET_REQ_COUNT_CMD.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I10779629be06275c844a2ad9afed72bd61c8a14a