Testing SPM and RMI together in parallel
The test randomly allocates either a SPM or RMI based test
onto a given CPU and does the same for all CPU's in the platform.
Once done, the test is run to determine if there are issues with
simultaneous execution of these tests.  The SPM tests are based
upon direct messaging and the RMI, granule delegation.

Signed-off-by: Mark Dykes <mark.dykes@arm.com>
Change-Id: Ida5da8852b5328187f7cab679bbe79b1bd7de790