ivy: shim: print welcome message after enabling the MMU
The shim prints a welcome message on boot. It uses the mp_printf()
function, which internally uses a spinlock. The shim currently
prints the welcome message *before* enabling the MMU, which causes
a hang while acquiring the spinlock.

This patch prints the welcome message *after* enabling the MMU to
fix the problem.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I8bd1ba2c4035fddf6f6a4c3d6844d69e7a11be7d