feat(WFxT): add a test for WFxT instructions
This patch adds the test to verify the WFET and WFIT instructions
introduced by FEAT_WFxT.

WFET and WFIT instructions assist in generating local-timeout event
and thereby act as wakeup event for the PE, when the virtual count
in CNTVCT_EL0 (counter-timer virtual count) register equals or exceed
the timeout value passed with these instructions.

Accordingly, this testcase verifies whether the timelapsed matches the
value supplied with WFET and WFIT instructions.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Signed-off-by: Jayanth Dodderi Chidanand <jayanthdodderi.chidanand@arm.com>
Change-Id: I9aea5da869db8520e305e49989cb71f166a582eb