test(interrupts): check if last service interrupt is Trusted Wdog timer
This patch adds checks to ensure the secure virtual interrupt signaled
to the SP by SPMC, when a trusted watchdog timer fired, was indeed
handled by Cactus SP.

It leverages the helper API introduced in the previous patches. The
robust checks done at the end of each test are no longer needed.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Ib0d6b63f8fdd180ef82c75c6d73a39883909e69e