test(cactus): prevent root region access from swd
This change adds TFTF and cactus test to permit checking a root region
cannot be accessed from secure world.
A hardcoded address marked Root in the GPT is shared to a secure
partition. The SP retrieves the region from the SPM, maps it and
attempts a read access to the region. It is expected to trigger a GPF
data abort on the PE caught by a custom exception handler.

Signed-off-by: Nabil Kahlouche <nabil.kahlouche@arm.com>
Change-Id: I882b53f784b8105e980d50014b74c4678b807567