test(spm): use ffa_helpers for ivy partition
To achieve this we create a common struct for all serivce calls
(smc, hvc and svc) we then use typedef to create specific type names
for each call to help with readability. We also implement a tftf_svc
function that uses the smccc_conduit so that both smc and svc calls
take the same type and can be easily swapped depending on the EL of
the SP. Finally we replace specific svc calls found in ivy_main.c with
the helper functions in ffa_helper.

Change-Id: Ic9525baabcf40d15545b6f6d504cf954373f08f9
Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>