refactor: mem share tests to avoid spurious error log
The test to request memory to be shared from SP to VM introduced a
spurious error message, as an SP is attempting a FFA_DIR_MSG_REQ to a
VM ID, which is not allowed in hafnium's implementation.

This patch is part of 2-part fix, refactoring the memory sharing tests:
- removed the 'words_to_write' field from CACTUS_MEM_SEND_CMD to use register to propagate the lender ID. The number of words to write is now
defined in a macro, as there is no hard requirement for it to be passed
in the commands arguments.
- Prior to this patch in handling CACTUS_REQ_MEM_SEND_CMD, the cactus SP
was also sending the CACTUS_MEM_SEND_CMD to the target of its memory
share call (which was causing the spurious error log). Instead, the
handling of CACTUS_REQ_MEM_SEND_CMD was changed to, in a successful
memory share operation, propagate the handle to tftf in the command
response which should then send to the borrower via CACTUS_MEM_SEND_CMD.

The tests should break with this patch.
The subsequent patch should reestablish order and conclude the fix.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I8ba6509e044ebfce618ae57985e7c720ed7e1967