fix(tftf): handle psci system suspend test failure
The TFTF system suspend test case failed with TRP randomly
on Newmore-CI. This failure is arbitary and hence could
not be reproduced locally.

However to prevent this failure in future, adding a small check
wherein the framework from now on will retry to turn on the
leadcpu for the next-tests if it fails in the first iteration.

Signed-off-by: Jayanth Dodderi Chidanand <jayanthdodderi.chidanand@arm.com>
Change-Id: I98bd5d8e2e81f1da6c03587db7c660a54171a910