test(interrupts): managed exit supported by first SP
This test exercises the following scenario:

NWd(tftf) sends a direct message request to the first SP which then
extends the call chain by sending direct message request to the second
SP. While the second SP is processing the direct message, a non-secure
interrupt gets triggered.

Managed exit is supported by the first SP but not supported by the
second SP in the call chain.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I872d36a5ae2996f6f7baccd698cc031405b6a109