test(interrupts): managed exit supported by second SP
This test exercises the following scenario:

NWd(tftf) sends a direct message request to the first SP which then
extends the call chain by sending direct message request to the second
SP. While the second SP is processing the direct message, a non-secure
interrupt gets triggered.

Managed exit is supported by the second SP but not supported by the
first SP in the call chain.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I5ee96e08be00a489720443c759c4ac24267b962b