test: exercise secure espi interrupt handling
This patch adds a test to trigger an espi interrupt when cactus is
running and ensure it is handled. Hafnium/SPMC added support for
enabling interrupts in the extended SPI range and this patch adds code
to map the GIC into cactus, so that the test can pend the espi interrupt
at the distributor and handle it.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: Ief932c40e3abd33d619f2b144e61cae449147b27