fix(build): clean el3 test payload image
The clean command also cleans the optional el3 test payload image even
if the platform does not support the el3_payload test. This will result
in makefile  error, since the el3_payload build system will try to
include the platform.mk file for the platform - which will be missing in
case of a unsupported platform. Hence, add a condition in the clean and
distclean command.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I27e1f3f66042b90ab1b0ee4780c587c6fd2a1f00