fix: serialize powering on secondary CPUs
This patch provides a temporary workaround for the following issue
seen with a notifications test in SPM test suite.

ERROR:   Unexpected affinity info state.
PANIC in EL3.

The reason for assertion fail in warmboot entrypath for a secondary
CPU needs further investigation but we believe it is possibly related
to the more common issue in the PSCI library as described here [1]

In order to increase the time between cache line invalidations,
TFTF brings secondary CPUs online one after the other.

[1] https://review.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/a291687de235da617afe25ad5226f448cac5c567/lib/psci/psci_on.c#122

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Ie523e40d4c90f52ecc8c7c42fce6ff214003ccc4