feat(rng_trap): modify tests for FEAT_RNG_TRAP
With the introduction of a handler for the trap generated upon an
attempt to read RNDR and RNDRRS registers, the old test will not
work anymore and needs to be changed so it does not longer expect
the system to panic but for EL3 return a valid random number after
handling the trap.

In addition, the test configs are unified as there is no need to
keep them separate since a system panic is not expected anymore.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>
Change-Id: I3ea4d09bbae58a906ea0be3b5448f5d3090571e7