Add "nop" test case
This is intended for use in conjunction with Trusted Firmware eXplorer
(TFX).  The dummy test case exists only to provide a symbol with a
well-known name ("debug_hook"), at which TFX can set a breakpoint and
therefore marshall all PEs on entry to Non-secure state.

The "nop" test case is not built by default, and therefore must be
explicitly included, e.g.

    make TESTS=nop

Signed-off-by: Gareth Stockwell <gareth.stockwell@arm.com>
Change-Id: I1d0f95714e5f662ca0f2d90db32020be15b43aed