test(interrupts): add tests for interrupt routing
This patch modifies existing secure interrupt tests to test interrupt
routing to particular cores/mpidr. Hafnium added a new field
"interrupts-target" to enable targeting interrupts to cores that are NOT
the boot core. This change modifies the cactus manifest to target the
secure watchdog timer (existing secure interrupt on fvp) to a particular
mpidr. To test interrupt handling, it converts the existing tests that
use the secure watchdog timer to test interrupt handling, to multi
core tests, such that the same test runs on the appropriate mpidr as
opposed to the boot core. Since the interrupt may be targeted to a
secondary PE, the corresponding secondary vcpus of cactus need a round
of ffa_run to function, which has also been added.
The mpidr to target to is platform specific and is defined as a platform
specific macro.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I27f48cde68945e59fb2ce84625ca145048eaffd1