refactor(ras): rename sdei related functions/events
Function to report sdei event was named on the basis of serror which has
caused that error. SError can be used in many cases and SDEI is one of
its usecases.

This patch also modifies inject_unrecoverable_ras_error() to accept wait
address as an argument to allow this to be used by other usecases.

Signed-off-by: Manish Pandey <manish.pandey2@arm.com>
Change-Id: I451008a7e5bfa34d8c4e8fb3c0c59b90c4bb7e5f