fix: memory region not page aligned
In [1] Hafnium mandates that every memory region allocated in the manifest must
be page aligned.

If the base address is not present, hafnium would allocate the memory region.
These memory regions are not guartanteed to be page aligned, and could conflict
with aforementioned patch.
The patch removes instances of such type of memory regions from cactus partition
manifest.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/17304

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I443e30b18531072cd108c58eb0762f3bffbaf65a