refactor(sve): move sve operations to a lib routine
This patch moves the SVE subtract operation to common sve library
routine and takes a callback function that does the world switch while
SVE operations are done in a loop.

This refactoring later helps to use this function to do context switch
from NS to Swd or from NS to Realm world based on the callback type.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: Iceb34b96fa85597be63a50c429ae0eb29f8fcaf8