feat(ff-a): partition information via registers
This patch enables basic support for getting partition information via
the ffa_partition_info_get_regs abi. This interface can be used to query
partition information in the absence of rx/tx buffer or when using
memory is inconvenient (such as early boot loaders etc). The patch adds
the required calls, a few helper functions and enables the use of x8-x17
as return values, that is required for this abi to work.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I70ed78e809a5bf77d77a49e5bc122c1989303ebb