fix(rme): reestalblish invalid access test
In [1] the invalid access test was removed from the respective
tests xml file, as it was leveraging a security vulnerability in
our SPM implementation.

As part of the same patchstack, we have harnessed and limited the
memory that an SP can use via manifest, or memory sharing operations.
This patch reestalblishes the invalid access that was discarded,
this time however only validating that the memory sharing of root
memory fails, which should be indicative that the SP can't access it.

[1] https://review.trustedfirmware.org/q/topic:%22ja%252Fmem_region_fix%22+(status:open%20OR%20status:merged)

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I6ee21586a2e94810fb4656054b26b2a51c97a544