feat(rmi): add support for FEAT_LPA2 to the Realm Extension tests
This patch adds support to the current Realm Extension tests to
enable and use 52 bit address size with 4KB granularity when
FEAT_LPA2 is present.

When the architecture supports LPA2 for 4KB granularity, this
patch assumes that RMM has support for it at stage 2 of address
translation.

This patch also removes any test for RMI extensions when no
building for aarch64 architecture.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I0ce4e5b26d8c87605baaca6f6948796f5716caa7