feat(xlat): add support for 52 bit PA size with 4KB granularity
This patch adds support to the xlat library to for 52Bits of
PA size with 4KB granularity (FEAT_LPA2). The patch only reports
the right granularity when it supports FEAT_LPA2 and it does
not enable the feature.

The patch also disables all the RMI tests when building for
Aarch32 architecture.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: Iee0cab8e1f8844a6af135771d8f940ec7e1dce84