fix(sme): use rdsvl instead of rdvl
Use rdsvl instruction to get Streaming SVE vector length instead of rdvl
instruction. When the CPU is in Streaming SVE mode both rdvl and rdsvl
instruction returns the same value but that is not true when the CPU is
in Normal SVE mode. So it's preferred to use rdsvl to get SVL.

This fix also aligns the buffer used for the ZA vector.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: Ieb6226f4fc275ee8a81eb359af465c26e307bc75