fix(fpu): write random value to fpsr/fpcr instead of read
The inline assembly code to write random values to FPSR and FPCR,
actually does a read. This zero out the random value that is used as a
template to fill FPU [Q0-Q31] registers.

When this helper routine is used by Non-secure, SP or Realm context
the FPU state written to the registers is always zero. The test cases
that verifies whether the FPU state is preserved across the world switch
might PASS always as it works with an empty FPU state.

This fix also changes fpu_state_print() to print the content of Q0-Q31
registers instead of the address.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I7bc2deba034b9bb4b6c1b15fe91f4562fd64d8f0