fix(tests/realm): enhance fpu state verification test
Select a random FPU FILL/CMP command to send to different security
states in a more effective way. If a FPU FILL command is issued to
Non-secure or Realm or Secure world, this change ensures that the next
command issued will be a FPU COMPARE command to a security state that is
different from the previous state.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I861892f2c340bb259eafa14d2fb540645ee5396a