test(smmu): secure peripheral access prevented
Check a secure peripheral upstream to an SMMU cannot perform
transactions to a realm region.

Add an NS region to cactus SP, to be used as a placeholder for the DMA
test engine to perform a RAND48 operation.
The test scenario changes the NS buffer PAS into Realm. It then queries
a SP to initiate a secure DMA operation on this buffer through the SMMU.
The operation is expected to fail as a secure DMA transaction to a Realm
region fails SMMU GPC checks.

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I5d061e5f0fc16cc9c14089f95e0ff5e49cb2a9cd