test(memory share): memory contents checks
In RME enabled systems, the memory is expected to be
scrubbed on PAS updates from S to NS. As well, it is likely
that the memory addresses are shadowed, and the contents are not
visible accross updates from the different address spaces.
As such, the SP shall not rely on memory content to be
in any form. FFA_MEM_LEND/FFA_MEM_DONATE are thus considered
for memory allocation purposes.

This patch changes the memory sharing tests, to cater
for the correct assumptions.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I78ed47df789dda1d695e2e52b0df8c28b12658da